Resolve "MUST-2 API must have a means of storing and processing events. Events must include location, the account that created them"

Closes #23

See merge request cis4250memeteam/poppin!10

+++ api/setup.sh
@@ -2,6 +2,16 @@
 
 export BASE_DIR=$(pwd)
 
+if [[ ! -f "${BASE_DIR}/setup.sh" ]];
+then
+  echo "~~~~~~"
+  echo "FAILED"
+  echo "~~~~~~"
+  echo "Please source setup.sh from the api/ directory"
+  printf "\tcd api && source setup.sh\n"
+  return 1
+fi
+
 cp git-hook-src/pre-push ../.git/hooks/pre-push
 
 python3 -m venv venv
